Muevo lógica de lectura y escritura del data.json a un archivo aparte.
