+++ ckanext/gobar_theme/routing.py
@@ -15,6 +15,7 @@ class GobArRouter:
         self.user_controller = 'ckanext.gobar_theme.user_controller:GobArUserController'
         self.google_analytics_controller = 'ckanext.gobar_theme.google_analytics_controller:GobArGAController'
         self.spatial_controller = 'ckanext.gobar_theme.spatial_controller:GobArSpatialController'
+        self.datajson_controller = 'ckanext.gobar_theme.lib.datajson_controller:GobArDatajsonController'
 
     def redirect(self, *routes):
         for url_from, url_to in routes:
@@ -218,7 +219,8 @@ class GobArRouter:
         )
 
     def connect_datajson(self):
-        self.home_routes.connect('datajson', '/data.json2', action='datajson')
+        with SubMapper(self.route_map, controller=self.datajson_controller) as m:
+            m.connect('datajson', '/data.json2', action='datajson')
         self.redirect(
             ('/datajson2', '/datajson2'),
         )
